fix expand, also in heredoc, add some malloc protects and frees..
