+++ main.c
@@ -6,7 +6,7 @@
 /*   By: eusatiko <eusatiko@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/21 15:07:50 by auspensk          #+#    #+#             */
-/*   Updated: 2024/09/11 13:47:33 by eusatiko         ###   ########.fr       */
+/*   Updated: 2024/09/13 11:50:25 by eusatiko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,6 @@ int	main(int argc, char *argv[], char *envp[])
 	data.cmd = parser(argv[1], &data);
 	return_value = execute_loop(&data);
 	if (data.cmd)
-		clean_exit(NULL, return_value, &data);
+		clean_exit(NULL, return_value, &data); //
 	clean_exit(NULL, 0, &data);
 }
