--- out_1
@@ -1,2 +1,0 @@
-test
-data->envp is 0x8722a0, data->tty_in is 0x8733a0, data->tty_out is (nil)
